Frontend

+++ src/screen/Cart.js
@@ -6,6 +6,7 @@ export default function Cart() {
   const [cartItem, setCartItem] = useState([]);
   const [removeSingleCartItem, setRemoveSingleCartItem] = useState([]);
   const [clearAllCart, setClearAllCartItem] = useState([]);
+  const [re, setRe] = useState(false);
 
   const token = localStorage.getItem("token");
   const user = token ? JSON.parse(token) : "";
@@ -14,7 +15,8 @@ export default function Cart() {
     axios
       .get(`http://localhost:5000/api/v1/shoppingcarts/cart-item/${userId}`)
       .then((res) => setCartItem(res.data));
-  },[]);
+    setRe(false);
+  }, [re]);
 
   const totalPrice = cartItem.reduce(
     (total, item) =>
@@ -40,13 +42,49 @@ export default function Cart() {
     return clearAllCart;
   };
 
-  const handleRemoveCartItem = async (cartId) => {
+  const handleRemoveCartItem = async (cartId, productId, proQty) => {
     try {
-      const response = await axios.delete(
-        `http://localhost:5000/api/v1/shoppingcarts/remove/cart_item/${cartId}`
+      // get the product data by product id
+      const productResponse = await axios.get(
+        `http://localhost:5000/api/v1/products/${productId}`
       );
-      setRemoveSingleCartItem(cartItem.filter((item) => item._id !== cartId));
-      return response;
+      const subStractCountInStock = productResponse.data;
+
+      // substract the countInStock of product by 1
+      subStractCountInStock.countInStock += proQty;
+
+      // get all the data of cart item by each user id
+      const response = await axios.get(
+        `http://localhost:5000/api/v1/shoppingcarts/cart-item/${userId}`
+      );
+      const items = response.data;
+
+      // check the exist cart item that is already exist
+      const existCartItem = items.find(
+        (item) => item.product._id === productId
+      );
+      if (existCartItem) {
+        existCartItem.quantity += proQty;
+        await axios.put(
+          `http://localhost:5000/api/v1/shoppingcarts/update-cart/${existCartItem._id}`,
+          {
+            quantity: existCartItem.quantity,
+          }
+        );
+
+        // implement the update of substract count_in_stock
+        await axios.put(
+          `http://localhost:5000/api/v1/products/update_count_in_stock/${productId}`,
+          subStractCountInStock
+        );
+
+        const response = await axios.delete(
+          `http://localhost:5000/api/v1/shoppingcarts/remove/cart_item/${cartId}`
+        );
+        setRemoveSingleCartItem(cartItem.filter((item) => item._id !== cartId));
+        setRe(true);
+        return response;
+      }
     } catch (err) {
       console.log(err);
     }
@@ -145,7 +183,13 @@ export default function Cart() {
                             <td className="cart__close">
                               <a
                                 href="#"
-                                onClick={() => handleRemoveCartItem(item.id)}
+                                onClick={() =>
+                                  handleRemoveCartItem(
+                                    item.id,
+                                    item.product.id,
+                                    item.quantity
+                                  )
+                                }
                               >
                                 <i className="fa fa-close" />
                               </a>
